Split batch API from american_option.hpp
The batch API (BatchAmericanOptionSolver, BatchAmericanOptionResult,
and solve_american_options_batch convenience functions) has been
extracted into a separate header file for better modularity.

This separation allows users who only need single-option pricing to
avoid including OpenMP dependencies. The batch API remains fully
functional with identical behavior and performance characteristics.

Changes:
- Created src/option/american_option_batch.hpp for batch API
- Removed batch code from src/option/american_option.hpp
- Added new american_option_batch build target with OpenMP flags
- Updated all consumers to include the new header
- Removed OpenMP linkopts from american_option target
- All tests pass without modification
