+++ benchmarks/BUILD.bazel
@@ -61,6 +61,7 @@ cc_binary(
     linkstatic = True,
     deps = [
         "//src/option:american_option",
+        "//src/option:american_option_batch",
         "//src/option:iv_solver_fdm",
         "//src/option:iv_solver_interpolated",
         "//src/option:normalized_chain_solver",
