+++ src/option/BUILD.bazel
@@ -31,8 +31,28 @@ cc_library(
         "-O3",
         "-march=native",      # Enable native CPU instructions (AVX-512 on modern CPUs)
         "-fopenmp-simd",      # Enable OpenMP SIMD pragmas for vectorization
-        "-fopenmp",           # Enable OpenMP for batch API parallelization
         "-ftree-vectorize",   # Enable auto-vectorization
+    ],
+    visibility = ["//visibility:public"],
+)
+
+cc_library(
+    name = "american_option_batch",
+    hdrs = ["american_option_batch.hpp"],
+    deps = [
+        ":american_option",
+        ":american_solver_workspace",
+        "//src/support:error_types",
+        "//src/support:parallel",
+    ],
+    copts = [
+        "-std=c++23",
+        "-Wall",
+        "-Wextra",
+        "-Werror",
+        "-O3",
+        "-march=native",
+        "-fopenmp",           # Enable OpenMP for batch API parallelization
     ],
     linkopts = ["-fopenmp"],
     visibility = ["//visibility:public"],
@@ -189,6 +209,7 @@ cc_library(
         ":option_spec",
         ":normalized_chain_solver",
         ":american_option",
+        ":american_option_batch",
         ":american_solver_workspace",
         ":price_table_snapshot_collector",
         "//src/support:error_types",
