+++ src/option/price_table_solver_factory.cpp
@@ -6,6 +6,7 @@
 #include "src/option/price_table_solver_factory.hpp"
 #include "src/option/normalized_chain_solver.hpp"
 #include "src/option/american_option.hpp"
+#include "src/option/american_option_batch.hpp"
 #include "src/option/american_solver_workspace.hpp"
 #include "src/option/price_table_snapshot_collector.hpp"
 #include "src/support/parallel.hpp"
