+++ tests/american_option_solver_test.cc
@@ -4,6 +4,7 @@
  */
 
 #include "src/option/american_option.hpp"
+#include "src/option/american_option_batch.hpp"
 #include "src/option/price_table_snapshot_collector.hpp"
 #include <gtest/gtest.h>
 #include <cmath>
