+++ tests/american_option_test.cc
@@ -1,4 +1,5 @@
 #include "src/option/american_option.hpp"
+#include "src/option/american_option_batch.hpp"
 
 #include <gtest/gtest.h>
 #include <cmath>
